requirements.txt
torch>=2.4
accelerate>=1.0
numpy>=1.26
pyyaml>=6.0.2
pytest>=8.3
packaging>=24.2
transformers>=4.57.0
